+++ src/types/mod.rs
@@ -5,4 +5,4 @@
 
 pub mod context;
 pub mod stardust;
-pub mod util;
+pub mod serde;
