+++ src/types/stardust/tangle/block/address/ed25519.rs
@@ -7,7 +7,7 @@ use iota_types::block::address as iota;
 use mongodb::bson::{spec::BinarySubtype, Binary, Bson};
 use serde::{Deserialize, Serialize};
 
-use crate::types::util::bytify;
+use crate::types::serde::bytify;
 
 /// A regular Ed25519 address.
 #[derive(Copy, Clone, Debug, PartialEq, Eq, Serialize, Deserialize, Hash)]
