+++ src/types/stardust/tangle/block/output/alias.rs
@@ -15,7 +15,7 @@ use super::{
     unlock_condition::{GovernorAddressUnlockCondition, StateControllerAddressUnlockCondition},
     OutputAmount, OutputId,
 };
-use crate::types::{context::TryFromWithContext, util::bytify};
+use crate::types::{context::TryFromWithContext, serde::bytify};
 
 /// Uniquely identifies an Alias.
 #[derive(Copy, Clone, Debug, PartialEq, Eq, Serialize, Deserialize, Hash)]
